requirements.txt
python-chess
requests
supabase
python-dotenv
psycopg2
pandas
beautifulsoup4